clamped player to screen bounds bumped project to cpp 17

+++ src/Entity.h
@@ -7,7 +7,7 @@ public:
 	Entity();
 
 	virtual void Input() {};
-	virtual void Update(double delta_time) {};
+	virtual void Update(float delta_time) {};
 	virtual void UpdateAnimation() {};
 	virtual void Draw() {};
